Add separate implementation for wide rhp context on native platforms
