+++ src/components/WideRHPContextProvider/index.tsx
@@ -8,35 +8,7 @@ import type {NavigationRoute} from '@libs/Navigation/types';
 import variables from '@styles/variables';
 import NAVIGATORS from '@src/NAVIGATORS';
 import SCREENS from '@src/SCREENS';
-
-type WideRHPContextType = {
-    // Route keys of screens that should be displayed in wide format
-    wideRHPRouteKeys: string[];
-
-    // Progress of changing format: 0 - narrow, 1 - wide
-    expandedRHPProgress: Animated.Value;
-
-    // Progress of the secondary overlay, the one covering wider RHP screen
-    secondOverlayProgress: Animated.Value;
-
-    // If the secondary overlay should be rendered. This value takes into account the delay of closing transition.
-    shouldRenderSecondaryOverlay: boolean;
-
-    // Show given route as in wide format
-    showWideRHPVersion: (route: NavigationRoute) => void;
-
-    // Remove given route from the array
-    cleanWideRHPRouteKey: (route: NavigationRoute) => void;
-
-    // Mark reportID as expense before condition check
-    markReportIDAsExpense: (reportID: string) => void;
-
-    // Check if reportID is marked as expense
-    isReportIDMarkedAsExpense: (reportID: string) => boolean;
-
-    // Navigate to the last element in wideRHPRouteKeys array
-    dismissToWideReport: () => void;
-};
+import type {WideRHPContextType} from './types';
 
 const expandedRHPProgress = new Animated.Value(0);
 const secondOverlayProgress = new Animated.Value(0);
@@ -249,5 +221,5 @@ function useShowWideRHPVersion(condition: boolean) {
 }
 
 export default WideRHPContextProvider;
-export type {WideRHPContextType};
+
 export {expandedRHPProgress, secondOverlayProgress, WideRHPContext, useShowWideRHPVersion, receiptPaneRHPWidth};
